old/requirements.txt
alembic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
sqlalchemy
flask_migrate
flask_script
jsonify
flask
wget
oauth2client
flask_sqlalchemy
google-cloud-storage
requests
uwsgi
queuelib
python-interface
boto3
gspread
kafka-python
jsonformatter
python3-logstash